+++ prime-dotnet-webapi/Services/Clients/ICollegeLicenceClient.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using Prime.Models;
+
+namespace Prime.Services.Clients
+{
+    public interface ICollegeLicenceClient
+    {
+        Task<PharmanetCollegeRecord> GetPharmanetCollegeRecordAsync(Certification certification);
+    }
+}
