--- prime-dotnet-webapi/Services/interfaces/IPharmanetApiService.cs
@@ -1,11 +1,0 @@
-using System.Threading.Tasks;
-
-using Prime.Models;
-
-namespace Prime.Services
-{
-    public interface IPharmanetApiService
-    {
-        Task<PharmanetCollegeRecord> GetCollegeRecordAsync(Certification certification);
-    }
-}
