+++ prime-dotnet-webapi/prime.csproj
@@ -6,37 +6,38 @@
     <NoWarn>$(NoWarn);1591</NoWarn>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="AspNetCore.HealthChecks.NpgSql" Version="3.0.0" />
-    <PackageReference Include="Microsoft.AspNet.WebApi.Client" Version="5.2.7" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.1.1" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.1" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.1" />
+    <PackageReference Include="AspNetCore.HealthChecks.NpgSql" Version="3.0.0"/>
+    <PackageReference Include="Microsoft.AspNet.WebApi.Client" Version="5.2.7"/>
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.1.1"/>
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.1"/>
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.1"/>
     <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="3.1.1">
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks.EntityFrameworkCore" Version="3.1.1" />
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="3.1.1" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="3.1.3" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.1" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="3.1.0" />
-    <PackageReference Include="Serilog.AspNetCore" Version="3.2.0" />
-    <PackageReference Include="Serilog.Enrichers.Environment" Version="2.1.3" />
-    <PackageReference Include="Serilog.Sinks.Async" Version="1.4.0" />
-    <PackageReference Include="SimpleBase" Version="3.0.0" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0" />
-    <PackageReference Include="Appccelerate.StateMachine" Version="5.1.0" />
-    <PackageReference Include="Swashbuckle.AspNetCore.Newtonsoft" Version="5.0.0" />
-    <PackageReference Include="CompareNETObjects" Version="4.65.0" />
+    <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks.EntityFrameworkCore" Version="3.1.1"/>
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="3.1.1"/>
+    <PackageReference Include="Microsoft.Extensions.Logging" Version="3.1.4"/>
+    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.1"/>
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.3"/>
+    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="3.1.0"/>
+    <PackageReference Include="Serilog.AspNetCore" Version="3.2.0"/>
+    <PackageReference Include="Serilog.Enrichers.Environment" Version="2.1.3"/>
+    <PackageReference Include="Serilog.Sinks.Async" Version="1.4.0"/>
+    <PackageReference Include="SimpleBase" Version="3.0.0"/>
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0"/>
+    <PackageReference Include="Appccelerate.StateMachine" Version="5.1.0"/>
+    <PackageReference Include="Swashbuckle.AspNetCore.Newtonsoft" Version="5.0.0"/>
+    <PackageReference Include="CompareNETObjects" Version="4.65.0"/>
+    <PackageReference Include="Microsoft.Extensions.Http" Version="3.1.4"/>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Configuration\Resources\ru-access-terms-v1.html" />
-    <EmbeddedResource Include="Configuration\Resources\ru-access-terms-v2.html" />
-    <EmbeddedResource Include="Configuration\Resources\ru-access-terms-v3.html" />
-    <EmbeddedResource Include="Configuration\Resources\obo-access-terms-v1.html" />
-    <EmbeddedResource Include="Configuration\Resources\obo-access-terms-v2.html" />
-    <EmbeddedResource Include="Configuration\Resources\obo-access-terms-v3.html" />
-    <EmbeddedResource Include="Configuration\Resources\obo-access-terms-v4.html" />
+    <EmbeddedResource Include="Configuration\Resources\ru-access-terms-v1.html"/>
+    <EmbeddedResource Include="Configuration\Resources\ru-access-terms-v2.html"/>
+    <EmbeddedResource Include="Configuration\Resources\ru-access-terms-v3.html"/>
+    <EmbeddedResource Include="Configuration\Resources\obo-access-terms-v1.html"/>
+    <EmbeddedResource Include="Configuration\Resources\obo-access-terms-v2.html"/>
+    <EmbeddedResource Include="Configuration\Resources\obo-access-terms-v3.html"/>
+    <EmbeddedResource Include="Configuration\Resources\obo-access-terms-v4.html"/>
   </ItemGroup>
 </Project>
